move Duration (incl. DSL) into scala.concurrent.duration package
so that the full package can be imported naturally:

import scala.concurrent.duration._

will give you all the types (Duration, FiniteDuration, Deadline) and the
DSL for constructing these.
