+++ src/actors-migration/scala/actors/Timeout.scala
@@ -8,7 +8,7 @@
 
 package scala.actors
 
-import scala.concurrent.util.Duration
+import scala.concurrent.duration.Duration
 import java.util.concurrent.TimeUnit
 import scala.language.implicitConversions
