--- src/library/scala/concurrent/util/duration/Classifier.scala
@@ -1,9 +1,0 @@
-package scala.concurrent.util.duration
-
-import scala.concurrent.util.{ FiniteDuration }
-
-trait Classifier[C] {
-  type R
-  def convert(d: FiniteDuration): R
-}
-
