+++ test/files/jvm/actmig-PinS_2.scala
@@ -1,5 +1,5 @@
 import scala.actors.{ MigrationSystem, StashingActor, ActorRef, Props, Exit }
-import scala.concurrent.util.duration._
+import scala.concurrent.duration._
 import scala.concurrent.{ Promise, Await }
 
 object SillyActor {
