+++ test/files/jvm/duration-tck.scala
@@ -2,8 +2,7 @@
  *  Copyright (C) 2012 Typesafe Inc. <http://www.typesafe.com>
  */
   
-import scala.concurrent.util._
-import duration._
+import scala.concurrent.duration._
 import scala.reflect._
 import java.util.concurrent.TimeUnit._
 import scala.tools.partest.TestUtil.intercept
